requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
discord.py==1.5.0
idna==2.10
multidict==4.7.6
peewee==3.13.3
python-dotenv==0.14.0
typing-extensions==3.7.4.3
yarl==1.5.1
